towhee/operators/resnet_image_embedding&towhee$main/requirements.txt
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.19.5
